同matlab进行对比

+++ src/algo/translate.h
@@ -11,7 +11,7 @@ extern "C"
 #endif
 
 	uint32_t translate_create(CModel *cm, uint32_t id, uint32_t dt);
-	uint32_t translate_setPar(CModel cm, float K, float T, float n, float Tao);
+	uint32_t translate_setPar(CModel cm, float K, uint32_t T, short n, float Tao);
 
 #ifdef __cplusplus
 }
